backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
httpx==0.27.2
aiohttp==3.11.7
feedparser==6.0.11
apscheduler==3.10.4
supabase==2.9.1
redis==5.2.0
pydantic==2.9.2
pydantic-settings==2.6.1
geopy==2.4.1
geojson==3.1.0
beautifulsoup4==4.12.3
lxml==5.3.0
groq==0.12.0
pytz==2024.2
python-multipart==0.0.17
starlette==0.41.2
websockets==14.1
